+++ MK4duo/src/planner/planner.h
@@ -32,20 +32,27 @@
 #ifndef PLANNER_H
 #define PLANNER_H
 
-class Planner;
-extern Planner planner;
-
-enum BlockFlag {
+enum BlockFlagBit {
   // Recalculate trapezoids on entry junction. For optimization.
-  BLOCK_FLAG_RECALCULATE          = _BV(0),
+  BLOCK_BIT_RECALCULATE,
 
   // Nominal speed always reached.
   // i.e., The segment is long enough, so the nominal speed is reachable if accelerating
   // from a safe speed (in consideration of jerking from zero speed).
-  BLOCK_FLAG_NOMINAL_LENGTH       = _BV(1),
+  BLOCK_BIT_NOMINAL_LENGTH,
 
   // Start from a halt at the start of this block, respecting the maximum allowed jerk.
-  BLOCK_FLAG_START_FROM_FULL_HALT = _BV(2)
+  BLOCK_BIT_START_FROM_FULL_HALT,
+
+  // The block is busy
+  BLOCK_BIT_BUSY
+};
+
+enum BlockFlag {
+  BLOCK_FLAG_RECALCULATE          = _BV(BLOCK_BIT_RECALCULATE),
+  BLOCK_FLAG_NOMINAL_LENGTH       = _BV(BLOCK_BIT_NOMINAL_LENGTH),
+  BLOCK_FLAG_START_FROM_FULL_HALT = _BV(BLOCK_BIT_START_FROM_FULL_HALT),
+  BLOCK_FLAG_BUSY                 = _BV(BLOCK_BIT_BUSY)
 };
 
 /**
@@ -59,31 +66,33 @@ enum BlockFlag {
  */
 typedef struct {
 
+  uint8_t flag;                             // Block flags (See BlockFlag enum above)
+
   unsigned char active_extruder;            // The extruder to move (if E move)
   unsigned char active_driver;              // Selects the active driver for E
 
   // Fields used by the bresenham algorithm for tracing the line
-  long steps[NUM_AXIS];                     // Step count along each axis
-  unsigned long step_event_count;           // The number of step events required to complete this block
+  int32_t steps[NUM_AXIS];                  // Step count along each axis
+  uint32_t step_event_count;                // The number of step events required to complete this block
 
   #if ENABLED(COLOR_MIXING_EXTRUDER)
-    unsigned long mix_event_count[MIXING_STEPPERS]; // Scaled step_event_count for the mixing steppers
+    uint32_t mix_event_count[MIXING_STEPPERS]; // Scaled step_event_count for the mixing steppers
   #endif
 
-  long accelerate_until,                    // The index of the step event on which to stop acceleration
-       decelerate_after,                    // The index of the step event on which to start decelerating
-       acceleration_rate;                   // The acceleration rate used for acceleration calculation
+  int32_t accelerate_until,                 // The index of the step event on which to stop acceleration
+          decelerate_after,                 // The index of the step event on which to start decelerating
+          acceleration_rate;                // The acceleration rate used for acceleration calculation
 
-  unsigned char direction_bits;             // The direction bit set for this block (refers to *_DIRECTION_BIT in config.h)
+  uint8_t direction_bits;                   // The direction bit set for this block (refers to *_DIRECTION_BIT in config.h)
 
   // Advance extrusion
   #if ENABLED(LIN_ADVANCE)
     bool use_advance_lead;
-    int e_speed_multiplier8; // Factorised by 2^8 to avoid float
+    uint32_t abs_adv_steps_multiplier8;     // Factorised by 2^8 to avoid float
   #elif ENABLED(ADVANCE)
-    long advance_rate;
-    volatile long initial_advance,
-                  final_advance;
+    int32_t advance_rate;
+    volatile int32_t  initial_advance,
+                      final_advance;
     float advance;
   #endif
 
@@ -93,7 +102,6 @@ typedef struct {
         max_entry_speed,                        // Maximum allowable junction entry speed in mm/sec
         millimeters,                            // The total travel of this block in mm
         acceleration;                           // acceleration mm/sec^2
-  uint8_t flag;                                 // Block flags (See BlockFlag enum above)
 
   // Settings for the trapezoid generator
   uint32_t nominal_rate,                        // The nominal step rate for this block in step_events/sec
@@ -101,25 +109,23 @@ typedef struct {
            final_rate,                          // The minimal rate at exit
            acceleration_steps_per_s2;           // acceleration steps/sec^2
 
-  unsigned long fan_speed;
+  uint32_t fan_speed;
 
   #if ENABLED(BARICUDA)
-    unsigned long valve_pressure, e_to_p_pressure;
+    uint32_t valve_pressure, e_to_p_pressure;
   #endif
 
   #if ENABLED(LASERBEAM)
     uint8_t laser_mode; // CONTINUOUS, PULSED, RASTER
     bool laser_status; // LASER_OFF, LASER_ON
     float laser_ppm; // pulses per millimeter, for pulsed and raster firing modes
-    unsigned long laser_duration; // laser firing duration in microseconds, for pulsed and raster firing modes
-    unsigned long steps_l; // step count between firings of the laser, for pulsed firing mode
+    uint32_t laser_duration; // laser firing duration in microseconds, for pulsed and raster firing modes
+    uint32_t steps_l; // step count between firings of the laser, for pulsed firing mode
     float laser_intensity; // Laser firing instensity in clock cycles for the PWM timer
     #if ENABLED(LASER_RASTER)
       unsigned char laser_raster_data[LASER_MAX_RASTER_LINE];
     #endif
   #endif 
-
-  volatile char busy;
 
 } block_t;
 
@@ -193,6 +199,11 @@ class Planner {
       static unsigned char old_direction_bits;
       // Segment times (in µs). Used for speed calculations
       static long axis_segment_time[2][3];
+    #endif
+
+    #if ENABLED(LIN_ADVANCE)
+      static float position_float[NUM_AXIS];
+      static float extruder_advance_k;
     #endif
 
     /**
@@ -368,7 +379,7 @@ class Planner {
     static block_t* get_current_block() {
       if (blocks_queued()) {
         block_t* block = &block_buffer[block_buffer_tail];
-        block->busy = true;
+        SBI(block->flag, BLOCK_BIT_BUSY);
         return block;
       }
       else
@@ -436,5 +447,7 @@ class Planner {
     static void recalculate();
 
 };
+
+extern Planner planner;
 
 #endif // PLANNER_H